feat: Add ScaleDetailsContext for fetching and managing scale details
- Implemented `ScaleDetailsProvider` to handle the context for scale details.
- Added `fetchScaleDetails` function to fetch and process scale data from the API.
- Introduced `getParams` utility function to parse URL parameters.
- Managed loading state and error handling for API requests.
- Updated local storage and session data based on API response.
- Provided context values for accessing scale details, report URLs, and loading state.
