+++ frontend/src/App.jsx
@@ -1,28 +1,27 @@
 import { Route, Routes } from "react-router-dom";
-import Healthcheck from "./Pages/SeverStatus/ServerHealthCheck";
-import Report from "./Pages/Reports/Report";
+import CreatingScale from "./Pages/CreateNewScale/CreatingScale";
+import EditScale from "./Pages/EditScale/EditScale";
+import HomePage from "./Pages/HomePage/HomePage";
+import LikertReport from "./Pages/LikertReport/LikertReport";
 import Login from "./Pages/Login/Login";
-import UserDetails from "./Pages/UserDetails/UserDetails";
+import Registration from './Pages/Registration/RegistrationPage';
+import OpenReportPage from "./Pages/Reports/OpenReportPage";
+import Report from "./Pages/Reports/Report";
 import ScaleDetails from "./Pages/Scale/ScaleDetails";
 import Scales from "./Pages/Scale/Scales";
-import OpenReportPage from "./Pages/Reports/OpenReportPage";
-import useDowellLogin from "./hooks/useDowellLogin";
-import HomePage from "./Pages/HomePage/HomePage";
-import Registration from './Pages/Registration/RegistrationPage';
-import LikertReport from "./Pages/LikertReport/LikertReport";
-import EditScale from "./Pages/EditScale/EditScale";
-import CreatingScale from "./Pages/CreateNewScale/CreatingScale";
-import Confirm from "./components/ScaleForm/Confirm";
 import ScalesReport from "./Pages/ScalesReport/ScalesReport";
+import Healthcheck from "./Pages/SeverStatus/ServerHealthCheck";
 import ShareScale from "./Pages/ShareScale/ShareScale";
+import UserDetails from "./Pages/UserDetails/UserDetails";
 import LLXLoginPage from "./Pages/llx/LLXLoginPage/LLXLoginPage";
-import LLXReport from "./Pages/llx/LLXReport/LLXReport";
-import LLXUserDetails from "./Pages/llx/LLXUserDetails/LLXUserDetails";
-import LLXScaleDetails from "./Pages/llx/LLXScaleDetails/LLXScaleDetails";
 import LLXOpenReportPage from "./Pages/llx/LLXOpenReportPage/LLXOpenReportPage";
-import LLXScale from "./Pages/llx/LLXScale/LLXScale";
 import NewLLXReport from "./Pages/llx/LLXReport/NewLLXReport";
+import LLXScale from "./Pages/llx/LLXScale/LLXScale";
 import NewLLXScaleDetails from "./Pages/llx/LLXScaleDetails/NewLLXScaleDetails";
+import LLXUserDetails from "./Pages/llx/LLXUserDetails/LLXUserDetails";
+import Confirm from "./components/ScaleForm/Confirm";
+import { ScaleDetailsProvider } from "./contexts/scaleDetailsContext";
+import useDowellLogin from "./hooks/useDowellLogin";
 
 const App = () => {
   useDowellLogin();
@@ -32,9 +31,13 @@ const App = () => {
       {/* login page only change the logo [done]*/}
       <Route path="/llx" element={<LLXLoginPage />} />
       {/* 49 number line [done]*/}
-      <Route path="/llx/reports" element={<NewLLXReport />} />
+      <Route path="/llx/reports" element={<ScaleDetailsProvider><NewLLXReport /></ScaleDetailsProvider>} />
       {/* <Route path="/llx/newreport" element={<NewLLXReport/>} /> */}
-      <Route path="/llx/scaledetails" element={<NewLLXScaleDetails/>} />
+      <Route path="/llx/scaledetails" element={
+  <ScaleDetailsProvider>
+    <NewLLXScaleDetails />
+  </ScaleDetailsProvider>
+} />
       
       {/* copy from ux live */}
       <Route path="/llx/report" element={<LLXOpenReportPage />} />
